service/agent2_design/requirements.txt
bedrock-agentcore
strands-agents
fastapi
boto3
pyyaml
bedrock-agentcore-starter-toolkit
strands-agents[otel]
requests
